+++ src/main/java/com/photostudio/web/servlet/order/DeleteOrderServlet.java
@@ -20,7 +20,7 @@ public class DeleteOrderServlet extends HttpServlet {
 
         String uri = request.getRequestURI();
         String[] partsOfUri = uri.split("/");
-        long orderId = Long.parseLong(partsOfUri[partsOfUri.length - 1]);
+        int orderId = Integer.parseInt(partsOfUri[partsOfUri.length - 1]);
 
         LOG.info("Request delete order is received: order {}", orderId);
